# docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: 电影票后端联调 API
  version: 0.2.0
  license:
    name: Proprietary
  description: |
    闲鱼插件、微信机器人、管理端与良票上游联调接口。

    鉴权分为三组：闲鱼接口使用 `X-Plugin-Token`，微信机器人接口使用
    `X-Bot-Token`，报价、订单和上游管理接口使用 `X-Admin-Token`。

    带有“高风险”字样的接口可能创建真实订单、提交上游或支付。手动测试前请确认
    `ticket.upstream.mock-enabled` 与 `ticket.xianyu.auto-fulfillment-enabled` 的配置。
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发环境
...
paths:
...
  /api/quotes:
    post:
      tags: [2. 报价管理]
      summary: 根据票务截图创建报价
      description: 可能调用真实 OCR、上传图片及询价，但不会创建订单或支付。
      operationId: createQuote
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
            example:
              imageBase64: '{{ticketImageBase64}}'
              customerId: apifox-customer-001
              channel: GOOFISH
      responses:
        '200':
          $ref: '#/components/responses/QuoteOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/quotes/{quoteNo}:
    get:
      tags: [2. 报价管理]
      summary: 查询报价
      operationId: getQuote
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/QuoteNo'
      responses:
        '200':
          $ref: '#/components/responses/QuoteOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/orders:
    post:
      tags: [3. 订单管理（高风险）]
      summary: 【高风险】根据报价创建并异步提交订单
      description: 创建本地订单后可能异步提交真实上游并支付。仅在明确确认真实出票后调用。
      operationId: createOrder
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              quoteNo: '{{quoteNo}}'
              customerId: apifox-customer-001
              paymentNo: apifox-payment-001
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/orders/{orderNo}:
    get:
      tags: [3. 订单管理（高风险）]
      summary: 查询本地订单
      operationId: getOrder
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
...
components:
  securitySchemes:
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: 本地 profile 默认值为 `local-admin-token`
    xianyuPluginToken:
      type: apiKey
      in: header
      name: X-Plugin-Token
      description: 用户控制台创建且已由管理员设置有效期的插件 Token；请求还必须携带 X-Plugin-Installation-Id
    botToken:
      type: apiKey
      in: header
      name: X-Bot-Token
      description: 用户控制台创建且已由管理员设置有效期的机器人 Token；请求还必须携带 X-Bot-Installation-Id
  parameters:
    QuoteNo:
      name: quoteNo
      in: path
      required: true
      schema:
        type: string
      example: Q20260516120000ABCDEFGH
    OrderNo:
      name: orderNo
      in: path
      required: true
      schema:
        type: string
      example: O20260516120000ABCDEFGH
...
  responses:
...
    QuoteOk:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/QuoteEnvelope'
    OrderOk:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OrderEnvelope'
...
    BadRequest:
      description: 参数校验或业务规则失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: imageBase64 is required
            data: null
    Unauthorized:
      description: Token 缺失或错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Unauthorized
            data: null
    AuthNotConfigured:
      description: 后端未配置对应接口 Token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: API token is not configured
            data: null
  schemas:
    ApiResponseBase:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: ok
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: object
              nullable: true
...
    QuoteEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/QuoteResponse'
    OrderEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/OrderResponse'
...
    SalesChannel:
      type: string
      enum: [WECHAT, GOOFISH]
    QuoteStatus:
      type: string
      enum: [CREATED, FAILED, ORDERED]
...
    OrderStatus:
      type: string
      enum: [CREATED, WAIT_SUBMIT, SUBMITTING, SUBMITTED, SUBMIT_FAILED, WAIT_PAY, PAID, TICKETING, ISSUED, FAILED, REFUNDED]
...
    CreateQuoteRequest:
      type: object
      required: [imageBase64]
      properties:
        imageBase64:
          type: string
          description: 图片 Base64 字符串；可包含或不包含 data URL 前缀，以后端实际配置为准
        customerId:
          type: string
        channel:
          $ref: '#/components/schemas/SalesChannel'
    MovieTicketInfo:
      type: object
      properties:
        taskId: {type: string}
        provinceName: {type: string}
        cityName: {type: string}
        areaName: {type: string}
        cityCode: {type: string}
        cinemaId: {type: string}
        cinemaCode: {type: string}
        cinemaAddress: {type: string}
        filmId: {type: string}
        filmImg: {type: string}
        customFilmType: {type: integer, format: int32}
        showId: {type: string}
        movieName: {type: string}
        cinemaName: {type: string}
        showTime: {type: string, format: date-time}
        hallName: {type: string}
        planType: {type: string}
        seatCount: {type: integer, format: int32}
        seats:
          type: array
          items: {type: string}
        seatsAndPrice:
          type: object
          additionalProperties: {type: string}
        maxPrice: {type: string}
        totalImagePrice: {type: string}
        imageUrl: {type: string}
    QuoteResponse:
      type: object
      properties:
        quoteNo:
          type: string
        ticketInfo:
          $ref: '#/components/schemas/MovieTicketInfo'
        upstreamPrice:
          type: number
        finalPrice:
          type: number
        totalPrice:
          type: number
        profit:
          type: number
        status:
          $ref: '#/components/schemas/QuoteStatus'
    CreateOrderRequest:
      type: object
      required: [quoteNo, customerId]
      properties:
        quoteNo:
          type: string
        customerId:
          type: string
        paymentNo:
          type: string
    OrderResponse:
      type: object
      properties:
        orderNo: {type: string}
        quoteNo: {type: string}
        customerId: {type: string}
        finalPrice: {type: number}
        totalPrice: {type: number}
        upstreamOrderId: {type: string, nullable: true}
        upstreamOrderNo: {type: string, nullable: true}
        upstreamOrderStatus: {type: integer, format: int32, nullable: true}
        ticketCodeInfo: {type: string, nullable: true}
        submitRetryCount: {type: integer, format: int32}
        lastSubmitError: {type: string, nullable: true}
        lastSyncError: {type: string, nullable: true}
        status:
          $ref: '#/components/schemas/OrderStatus'
        statusText: {type: string}
        terminal: {type: boolean}
        shouldPoll: {type: boolean}
...